pyproject.toml
[tool.poetry]
name = "repo-insights"
version = "0.1.0"
description = "Tool to get insights out of git data from repos"
authors = ["Juan Abia <juan.abia999@gmail.com>"]
readme = "README.md"
packages = [{include = "repo_insights"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pyyaml = "^6.0.1"
prettytable = "^3.9.0"
plotly = "^5.17.0"
pandas = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
